docker/lwua-ingest/lwua-py/pyproject.toml
[tool.poetry]
name = "lwua"
version = "0.0.0"
description = "lifewatch user analysis"
authors = ["Laurian Van Maldeghem <laurian.van.maldeghem@vliz.be>", "Cedric Decruw <cedric.decruw@vliz.be>", "Marc Portier <marc.portier@vliz.be>"]
license = "CC0"

[tool.poetry.dependencies]
python = "^3.10"
apscheduler = "^3.10.4"
pyaml = "^23.9.3"
python-dotenv = "^1.0.0"
sparqlwrapper = "^2.0.0"
pyrdfj2 = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
